core/ledger/zkp_verification.py
import hashlib
import random

class ZKPVerification:
    def __init__(self):
        """
        Initializes the ZeroKnowledgeProof (ZKP) system for verifying computations without revealing inputs.
        """
        pass

    def generate_proof(self, secret: int, base: int, modulus: int) -> dict:
        """
        Generates a proof for the knowledge of a secret using a simple Schnorr-like protocol.
        :param secret: The secret value.
        :param base: The base value used in modular exponentiation.
        :param modulus: The modulus value used in modular exponentiation.
        :return: A dictionary containing the generated proof details.
        """
        commitment = pow(base, secret, modulus)  # Compute the commitment
        random_value = random.randint(1, modulus - 1)  # Random value for blinding
        blinded_commitment = pow(base, random_value, modulus)  # Blinded commitment

        # Hashing the blinded commitment to create the challenge
        challenge = int(hashlib.sha256(str(blinded_commitment).encode()).hexdigest(), 16) % modulus

        # Response calculation
        response = (random_value + secret * challenge) % modulus

        proof = {
            "commitment": commitment,
            "blinded_commitment": blinded_commitment,
            "challenge": challenge,
            "response": response
        }

        print(f"Generated proof: {proof}")
        return proof

    def verify_proof(self, proof: dict, base: int, modulus: int) -> bool:
        """
        Verifies a proof for the knowledge of a secret without revealing the secret itself.
        :param proof: The proof details containing commitment, challenge, and response.
        :param base: The base value used in modular exponentiation.
        :param modulus: The modulus value used in modular exponentiation.
        :return: True if the proof is valid, False otherwise.
        """
        # Recalculate the expected commitment using the response and challenge
        left_hand_side = pow(base, proof["response"], modulus)
        right_hand_side = (proof["blinded_commitment"] * pow(proof["commitment"], proof["challenge"], modulus)) % modulus

        is_valid = left_hand_side == right_hand_side
        if is_valid:
            print("Proof verified successfully.")
        else:
            print("Proof verification failed.")
        return is_valid
    
    def validate_agent_computation(self, agent_id, computation_data):
        # Method logic here
        pass

# Example usage
if __name__ == "__main__":
    zkp = ZKPVerification()

    # Set parameters for ZKP
    secret = 42
    base = 5
    modulus = 97

    # Generate proof of knowledge of the secret
    proof = zkp.generate_proof(secret, base, modulus)

    # Verify the generated proof
    verification_result = zkp.verify_proof(proof, base, modulus)
    print(f"Verification result: {verification_result}")
